Adds scripts to generate delta f coeffs, also includes updates to pdg arrays for larger number or resonances

+++ src/cpp/readindata.h
@@ -51,7 +51,7 @@ class Plasma
 
 typedef struct
 {
-  int mc_id; // Monte Carlo number according PDG
+  long int mc_id; // Monte Carlo number according PDG
   string name;
   double mass;
   double width;
